Enables passing of function to `appendTo`
Instead of assigning an Element to `appendTo`,
one can now pass a function that evaluates to an Element

The appendTo function is now evaluated during createToolTips/initialization
of the tippy instance, which is good enough for not leaking appendTo
when reinitializing for various DOM changes e.g. in between
multiple `turbolinks:load` transitions.

This also changes the Defaults.appendTo into a function reevaluating the ‘current’ document.body